ARMS-149 #Done Make validation error message default
reports.create.report.validation.message is the id used in the
form-components/form-template.vm to show user there is a validition error.
The language file provide the content of this id is reports.properties. With
region-aware language service, it needs to make this id:content pair in the
default sets of language files.